Rename: 작성 코드 answer 폴더로 통합 - answer 폴더로 통합시 context null 에러가 발생해서 ver2 코드는 전부 주석처리 해둔 상태입니다.
