+++ app/src/pages/DarkMode/answer/ver1/components/Header.jsx
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import S from "../../DarkMode/DarkMode.module.scss";
+import S from "../DarkMode.module.scss";
 
 const Header = ({ theme }) => {
   return (
