+++ app/src/pages/DarkMode/answer/ver1/index.jsx
@@ -2,9 +2,6 @@ import { useState } from "react";
 
 import Page from "./ver1/components/Page";
 
-// 최상위 main.jsx 컴포넌트라고 가정합니다.
-// 먼저 props state와 props로 구현하고
-// 이후 contextAPI를 사용해서 props drilling를 해결해봅니다.
 const DarkMode = () => {
   const [theme, setTheme] = useState(false);
