+++ app/src/pages/DarkMode/answer/ver2/DarkMode.module.scss
@@ -1,0 +1,100 @@
+$active-color: #6fb994;
+
+.page {
+  width: 100%;
+  height: 80vh;
+  display: flex;
+  flex-direction: column;
+}
+
+.header {
+  width: 100%;
+  height: 80px;
+  border-bottom: 1px solid rgb(210, 210, 210);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.header .headerTitle {
+  font-size: 20px;
+  font-weight: 600;
+  padding-bottom: 0;
+}
+
+.content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  font-size: 30px;
+}
+
+.content .title {
+  font-size: 40px;
+  font-weight: 600;
+  margin-bottom: 0;
+}
+
+.footer {
+  width: 100%;
+  height: 80px;
+  border-top: 1px solid rgb(210, 210, 210);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.footer .name {
+  font-size: 14px;
+  margin-bottom: 0;
+}
+
+.switch {
+  position: relative;
+  margin: 24px 0 40px 0;
+  .themeInput[type="checkbox"] {
+    visibility: hidden;
+    &:checked {
+      + label {
+        background-color: $active-color;
+        &::after {
+          left: 27px;
+        }
+      }
+    }
+  }
+
+  .themeLabel {
+    position: absolute;
+    width: 50px;
+    height: 23px;
+    top: 0;
+    left: 0;
+    background-color: #b4c1ba;
+    border-radius: 50px;
+    &::after {
+      content: "";
+      width: 25px;
+      height: 25px;
+      background-color: #478e6a;
+      position: absolute;
+      left: -1px;
+      top: -1px;
+      border-radius: 50%;
+      box-shadow: 0 0 4px #999;
+      cursor: pointer;
+    }
+  }
+}
+
+.light {
+  color: #222;
+  background: #f7f7f7;
+}
+
+.dark {
+  color: #fff;
+  background: rgb(23, 32, 42);
+}
