+++ app/src/pages/DarkMode/answer/ver2/components/Page.jsx
@@ -1,0 +1,16 @@
+import Header from "./Header";
+import Content from "./Content";
+import Footer from "./Footer";
+import S from "../DarkMode.module.scss";
+
+const Page = () => {
+  return (
+    <div className={S.page}>
+      <Header />
+      <Content />
+      <Footer />
+    </div>
+  );
+};
+
+export default Page;
