+++ app/src/pages/DarkMode/answer/ver2/context/themeContext.jsx
@@ -1,0 +1,3 @@
+// import { createContext } from "react";
+
+// export const ThemeContext = createContext(null);
